feat: Implementa regras de negócio na camada de Serviço

+++ src/model/services/ClasseService.java
@@ -2,14 +2,18 @@ package model.services;
 
 import java.util.List;
 
+import db.DbException;
 import model.dao.ClasseDao;
+import model.dao.PersonagemDao;
 import model.entities.Classe;
 	
 public class ClasseService {
 	private ClasseDao dao;
+	private PersonagemDao personagemDao;
 	
-	public ClasseService(ClasseDao dao) {
+	public ClasseService(ClasseDao dao, PersonagemDao personagemDao) {
 		this.dao = dao;
+		this.personagemDao = personagemDao;
 	}
 
 	public List<Classe> findAll(){
@@ -17,7 +21,7 @@ public class ClasseService {
 	}
 	
 	public void saveOrUpdate(Classe obj) {
-        if (obj.getNome() == null || obj.getNome().trim().isEmpty()) { //exemplo de rn
+        if (obj.getNome() == null || obj.getNome().trim().isEmpty()) { 
             System.out.println("Erro de validação: O nome da classe não pode ser vazio.");
             return; 
         }
@@ -30,6 +34,10 @@ public class ClasseService {
 	}
 	
 	public void remove(Classe obj) {
+		
+		if (personagemDao.countByRacaId(obj.getId()) > 0) {
+            throw new DbException("Não é possível deletar esta raça, pois ela está sendo utilizada por personagens.");
+        }
         dao.deleteById(obj.getId());
     }
 }
